tests: init
