+++ package.json
@@ -7,8 +7,8 @@
     "create-html5-boilerplate": "./index.js"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "try": "node ./index.js"
+    "test": "jest",
+    "start": "node ./index.js"
   },
   "keywords": [],
   "author": "Vlad Tansky",
@@ -20,5 +20,8 @@
     "ora": "^4.0.4",
     "pacote": "^11.1.10",
     "yargs-parser": "^18.1.3"
+  },
+  "devDependencies": {
+    "jest": "^26.0.1"
   }
 }
